refactor(bin): move command setup/execution code into subfolder
I don't want all of this state hanging around main.

I've had to declare many variables as static, which is usually bad.
I need to pass data from CLI configure time to the post-parse callback.
The underlying storage needs to have the same lifetime as the CLI::App.
This means the storage must be in main, or a global.

The point of this refactor was to move logic out of main, so globals are the lesser evil.
Just don't try and spawn multiple CLI instances at once.
