+++ bin/commands/license.cpp
