+++ bin/commands/ls.cpp
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2023 J. Stanley Warford, Matthew McRaven
- *
+ * Copyright (c) 2023-2024 J. Stanley Warford, Matthew McRaven
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -16,9 +15,9 @@
  */
 
 #include "ls.hpp"
-#include "./shared.hpp"
-#include "help/builtins/figure.hpp"
 #include <iostream>
+#include "../shared.hpp"
+#include "help/builtins/figure.hpp"
 
 ListTask::ListTask(int ed, QObject *parent) : Task(parent), ed(ed) {}
 
@@ -33,22 +32,18 @@ void ListTask::run() {
   // Prevent figure name from overlapping with file types. See #305.
   int maxFigWidth = 10;
   for (auto figure : figures)
-    maxFigWidth = std::max<int>(figure->chapterName().length() + 1 /*.*/ +
-                                    figure->figureName().length(),
-                                maxFigWidth);
+    maxFigWidth = std::max<int>(figure->chapterName().length() + 1 /*.*/ + figure->figureName().length(), maxFigWidth);
   for (auto problem : problems)
-    maxFigWidth = std::max<int>(problem->chapterName().length() + 1 /*.*/ +
-                                    problem->figureName().length(),
-                                maxFigWidth);
+    maxFigWidth =
+        std::max<int>(problem->chapterName().length() + 1 /*.*/ + problem->figureName().length(), maxFigWidth);
 
   int maxMacroWidth = 6;
   for (auto macro : macros)
     maxMacroWidth = std::max<int>(macro->name().length(), maxMacroWidth);
   std::cout << "Figures: " << std::endl;
   for (auto &figure : figures) {
-    std::cout << u"%1.%2"_qs.arg(figure->chapterName(), figure->figureName())
-                     .leftJustified(maxFigWidth + 2)
-                     .toStdString();
+    std::cout
+        << u"%1.%2"_qs.arg(figure->chapterName(), figure->figureName()).leftJustified(maxFigWidth + 2).toStdString();
     std::cout << figure->typesafeElements().keys().join(", ").toStdString();
     std::cout << std::endl;
   }
@@ -56,8 +51,7 @@ void ListTask::run() {
   if (problems.size() > 0) {
     std::cout << "\nProblems: \n";
     for (auto &problem : problems) {
-      std::cout << u"%1.%2"_qs
-                       .arg(problem->chapterName(), problem->figureName())
+      std::cout << u"%1.%2"_qs.arg(problem->chapterName(), problem->figureName())
                        .leftJustified(maxFigWidth + 2)
                        .toStdString();
       std::cout << problem->typesafeElements().keys().join(", ").toStdString();
@@ -68,9 +62,7 @@ void ListTask::run() {
   if (macros.size() > 0) {
     std::cout << "\nMacros: \n";
     for (auto &macro : macros)
-      std::cout << u"%1"_qs.arg(macro->name())
-                       .leftJustified(maxMacroWidth + 2)
-                       .toStdString()
+      std::cout << u"%1"_qs.arg(macro->name()).leftJustified(maxMacroWidth + 2).toStdString()
                 << u"%1"_qs.arg(macro->argCount()).toStdString() << std::endl;
   }
   return emit finished(0);
